@@ -12,18 +12,16 @@ import marko
 import marko.element
 import marko.inline
 from docling_core.types.doc import (
-    ContentLayer,
-    DocItem,
     DocItemLabel,
     DoclingDocument,
     DocumentOrigin,
+    Formatting,
     ListItem,
     NodeItem,
     TableCell,
     TableData,
     TextItem,
 )
-from docling_core.types.doc.document import Formatting
 from marko import Markdown
 from pydantic import AnyUrl, BaseModel, Field, TypeAdapter
 from typing_extensions import Annotated, override
@@ -73,6 +71,21 @@ _CreationPayload = Annotated[
 
 
 class MarkdownDocumentBackend(DeclarativeDocumentBackend):
+    _ENTITY_RE = re.compile(r"&(#\d+|#x[0-9a-fA-F]+|\w+);")
+
+    @staticmethod
+    def _unescape_except_pipe(text: str) -> str:
+        def replace(match):
+            entity = match.group(0)
+
+            # entities that represent |
+            if entity in ("&#124;", "&#x7C;", "&vert;"):
+                return entity
+
+            return unescape(entity)
+
+        return MarkdownDocumentBackend._ENTITY_RE.sub(replace, text)
+
     def _shorten_underscore_sequences(self, markdown_text: str, max_length: int = 10):
         # This regex will match any sequence of underscores
         pattern = r"_+"
@@ -171,7 +184,7 @@ class MarkdownDocumentBackend(DeclarativeDocumentBackend):
                         1  # currently supporting just simple tables (without spans)
                     )
                     icell = TableCell(
-                        text=cellval.strip(),
+                        text=unescape(cellval.strip()),
                         row_span=row_span,
                         col_span=col_span,
                         start_row_offset_idx=trow_ind,
@@ -365,21 +378,26 @@ class MarkdownDocumentBackend(DeclarativeDocumentBackend):
                 element.dest
             )
 
-        elif isinstance(element, (marko.inline.RawText, marko.inline.Literal)):
+        elif isinstance(element, marko.inline.RawText | marko.inline.Literal):
             _log.debug(f" - RawText/Literal: {element.children}")
-            snippet_text = (
-                element.children.strip() if isinstance(element.children, str) else ""
+            original_text = (
+                element.children if isinstance(element.children, str) else ""
             )
-            snippet_text = unescape(snippet_text)
-            # Detect start of the table:
-            if "|" in snippet_text or self.in_table:
-                # most likely part of the markdown table
+            snippet_text = unescape(original_text.strip())
+            is_table_row = "|" in snippet_text and (
+                self.in_table or original_text.lstrip().startswith("|")
+            )
+            if is_table_row:
                 self.in_table = True
-                if len(self.md_table_buffer) > 0:
+            if self.in_table and snippet_text:
+                snippet_text = self._unescape_except_pipe(original_text.strip())
+                # If we're in a table, keep adding text (for formatted content in cells)
+                if self.md_table_buffer:
                     self.md_table_buffer[len(self.md_table_buffer) - 1] += snippet_text
                 else:
                     self.md_table_buffer.append(snippet_text)
             elif snippet_text:
+                # Not in table - close any pending table and process as regular text
                 self._close_table(doc)
 
                 if creation_stack:
@@ -440,7 +458,7 @@ class MarkdownDocumentBackend(DeclarativeDocumentBackend):
             )
 
         elif (
-            isinstance(element, (marko.block.CodeBlock, marko.block.FencedCode))
+            isinstance(element, marko.block.CodeBlock | marko.block.FencedCode)
             and len(element.children) > 0
             and isinstance((child := element.children[0]), marko.inline.RawText)
             and len(snippet_text := (child.children.strip())) > 0
@@ -482,7 +500,7 @@ class MarkdownDocumentBackend(DeclarativeDocumentBackend):
                 _log.debug(f"Some other element: {element}")
 
         if (
-            isinstance(element, (marko.block.Paragraph, marko.block.Heading))
+            isinstance(element, marko.block.Paragraph | marko.block.Heading)
             and len(element.children) > 1
         ):
             parent_item = doc.add_inline_group(parent=parent_item)
