@@ -37,7 +37,7 @@ GRANITEDOCLING_TRANSFORMERS = InlineVlmOptions(
     stop_strings=["</doctag>", "<|end_of_text|>"],
 )
 
-GRANITEDOCLING_VLLM = GRANITEDOCLING_TRANSFORMERS.model_copy()
+GRANITEDOCLING_VLLM = GRANITEDOCLING_TRANSFORMERS.model_copy(deep=True)
 GRANITEDOCLING_VLLM.inference_framework = InferenceFramework.VLLM
 
 GRANITEDOCLING_MLX = InlineVlmOptions(
@@ -68,7 +68,7 @@ GRANITEDOCLING_VLLM_API = ApiVlmOptions(
     response_format=ResponseFormat.DOCTAGS,
 )
 
-GRANITEDOCLING_OLLAMA = GRANITEDOCLING_VLLM_API.model_copy()
+GRANITEDOCLING_OLLAMA = GRANITEDOCLING_VLLM_API.model_copy(deep=True)
 GRANITEDOCLING_OLLAMA.url = AnyUrl("http://localhost:11434/v1/chat/completions")
 GRANITEDOCLING_OLLAMA.params["model"] = "ibm/granite-docling:258m"
 
@@ -317,6 +317,22 @@ DOLPHIN_TRANSFORMERS = InlineVlmOptions(
     temperature=0.0,
 )
 
+# DeepSeek-OCR
+DEEPSEEKOCR_OLLAMA = ApiVlmOptions(
+    url="http://localhost:11434/v1/chat/completions",
+    params=dict(
+        model="deepseek-ocr:3b",
+        max_tokens=4096,
+        skip_special_tokens=True,
+    ),
+    prompt="<|grounding|>Convert the document to markdown. ",
+    timeout=90,
+    scale=2.0,
+    temperature=0.0,
+    concurrency=4,
+    response_format=ResponseFormat.DEEPSEEKOCR_MARKDOWN,
+)
+
 # NuExtract
 NU_EXTRACT_2B_TRANSFORMERS = InlineVlmOptions(
     repo_id="numind/NuExtract-2.0-2B",
@@ -346,3 +362,4 @@ class VlmModelType(str, Enum):
     GOT_OCR_2 = "got_ocr_2"
     GRANITEDOCLING = "granite_docling"
     GRANITEDOCLING_VLLM = "granite_docling_vllm"
+    DEEPSEEKOCR_OLLAMA = "deepseekocr_ollama"
