@@ -7,12 +7,11 @@ from docling_core.types.doc import (
     DescriptionMetaField,
     DoclingDocument,
     NodeItem,
+    PictureClassificationLabel,
     PictureItem,
     PictureMeta,
 )
-from docling_core.types.doc.document import (  # TODO: move import to docling_core.types.doc
-    PictureDescriptionData,
-)
+from docling_core.types.doc.document import PictureDescriptionData
 from PIL import Image
 
 from docling.datamodel.accelerator_options import AcceleratorOptions
@@ -75,6 +74,13 @@ class PictureDescriptionBaseModel(
                         area_fraction = prov.bbox.area() / page_area
                         if area_fraction < self.options.picture_area_threshold:
                             describe_image = False
+            if describe_image and not _passes_classification(
+                el.item.meta,
+                self.options.classification_allow,
+                self.options.classification_deny,
+                self.options.classification_min_confidence,
+            ):
+                describe_image = False
             if describe_image:
                 elements.append(el.item)
                 images.append(el.image)
@@ -101,3 +107,43 @@ class PictureDescriptionBaseModel(
     @abstractmethod
     def get_options_type(cls) -> Type[PictureDescriptionBaseOptions]:
         pass
+
+
+def _passes_classification(
+    meta: Optional[PictureMeta],
+    allow: Optional[List[PictureClassificationLabel]],
+    deny: Optional[List[PictureClassificationLabel]],
+    min_confidence: float,
+) -> bool:
+    if not allow and not deny:
+        return True
+    predicted = None
+    if meta and meta.classification:
+        predicted = meta.classification.predictions
+    if not predicted:
+        return allow is None
+    if deny:
+        deny_set = {_label_value(label) for label in deny}
+        for entry in predicted:
+            if _meets_confidence(entry.confidence, min_confidence) and (
+                entry.class_name in deny_set
+            ):
+                return False
+    if allow:
+        allow_set = {_label_value(label) for label in allow}
+        return any(
+            _meets_confidence(entry.confidence, min_confidence)
+            and entry.class_name in allow_set
+            for entry in predicted
+        )
+    return True
+
+
+def _label_value(label: Union[PictureClassificationLabel, str]) -> str:
+    return label.value if isinstance(label, PictureClassificationLabel) else str(label)
+
+
+def _meets_confidence(confidence: Optional[float], min_confidence: float) -> bool:
+    return min_confidence <= 0 or (
+        confidence is not None and confidence >= min_confidence
+    )
