@@ -1,10 +1,10 @@
 def ocr_engines():
-    from docling.models.auto_ocr_model import OcrAutoModel
-    from docling.models.easyocr_model import EasyOcrModel
-    from docling.models.ocr_mac_model import OcrMacModel
-    from docling.models.rapid_ocr_model import RapidOcrModel
-    from docling.models.tesseract_ocr_cli_model import TesseractOcrCliModel
-    from docling.models.tesseract_ocr_model import TesseractOcrModel
+    from docling.models.stages.ocr.auto_ocr_model import OcrAutoModel
+    from docling.models.stages.ocr.easyocr_model import EasyOcrModel
+    from docling.models.stages.ocr.ocr_mac_model import OcrMacModel
+    from docling.models.stages.ocr.rapid_ocr_model import RapidOcrModel
+    from docling.models.stages.ocr.tesseract_ocr_cli_model import TesseractOcrCliModel
+    from docling.models.stages.ocr.tesseract_ocr_model import TesseractOcrModel
 
     return {
         "ocr_engines": [
@@ -19,8 +19,12 @@ def ocr_engines():
 
 
 def picture_description():
-    from docling.models.picture_description_api_model import PictureDescriptionApiModel
-    from docling.models.picture_description_vlm_model import PictureDescriptionVlmModel
+    from docling.models.stages.picture_description.picture_description_api_model import (
+        PictureDescriptionApiModel,
+    )
+    from docling.models.stages.picture_description.picture_description_vlm_model import (
+        PictureDescriptionVlmModel,
+    )
 
     return {
         "picture_description": [
@@ -34,7 +38,7 @@ def layout_engines():
     from docling.experimental.models.table_crops_layout_model import (
         TableCropsLayoutModel,
     )
-    from docling.models.layout_model import LayoutModel
+    from docling.models.stages.layout.layout_model import LayoutModel
 
     return {
         "layout_engines": [
@@ -45,7 +49,9 @@ def layout_engines():
 
 
 def table_structure_engines():
-    from docling.models.table_structure_model import TableStructureModel
+    from docling.models.stages.table_structure.table_structure_model import (
+        TableStructureModel,
+    )
 
     return {
         "table_structure_engines": [
