@@ -17,7 +17,7 @@ from docling.datamodel.base_models import (
 )
 from docling.datamodel.document import ConversionResult
 from docling.models.base_model import BasePageModel
-from docling.models.layout_model import LayoutModel
+from docling.models.stages.layout.layout_model import LayoutModel
 from docling.utils.profiling import TimeRecorder
 
 _log = logging.getLogger(__name__)
