@@ -15,13 +15,20 @@ from docling.datamodel.vlm_model_specs import (
     SMOLDOCLING_MLX,
     SMOLDOCLING_TRANSFORMERS,
 )
-from docling.models.code_formula_model import CodeFormulaModel
-from docling.models.document_picture_classifier import DocumentPictureClassifier
-from docling.models.easyocr_model import EasyOcrModel
-from docling.models.layout_model import LayoutModel
-from docling.models.picture_description_vlm_model import PictureDescriptionVlmModel
-from docling.models.rapid_ocr_model import RapidOcrModel
-from docling.models.table_structure_model import TableStructureModel
+from docling.models.stages.code_formula.code_formula_model import CodeFormulaModel
+from docling.models.stages.layout.layout_model import LayoutModel
+from docling.models.stages.ocr.easyocr_model import EasyOcrModel
+from docling.models.stages.ocr.rapid_ocr_model import RapidOcrModel
+from docling.models.stages.picture_classifier.document_picture_classifier import (
+    DocumentPictureClassifier,
+    DocumentPictureClassifierOptions,
+)
+from docling.models.stages.picture_description.picture_description_vlm_model import (
+    PictureDescriptionVlmModel,
+)
+from docling.models.stages.table_structure.table_structure_model import (
+    TableStructureModel,
+)
 from docling.models.utils.hf_model_download import download_hf_model
 
 _log = logging.getLogger(__name__)
@@ -69,8 +76,11 @@ def download_models(
 
     if with_picture_classifier:
         _log.info("Downloading picture classifier model...")
+        pic_opts = DocumentPictureClassifierOptions()
         DocumentPictureClassifier.download_models(
-            local_dir=output_dir / DocumentPictureClassifier._model_repo_folder,
+            repo_id=pic_opts.repo_id,
+            revision=pic_opts.revision,
+            local_dir=output_dir / pic_opts.repo_cache_folder,
             force=force,
             progress=progress,
         )
